Start to separate schema files into seperate extensions
This allows the collision primitives to be used for use-cases which
don't require simulation.

Temp names for now while I experiment with how assets feel. Side-effect that
material had to move from a collider property to a node
extension, i.e., instead of:

"extensions" : {
    "MSFT_Physics" : {
        "collider" : {
            "box" : { "size" : [ 2, 2, 2 ] },
            "physicsMaterial" : 1
        }
    }
}

We now specify the same thing via:

"extensions" : {
    "MSFT_CollisionPrimitives" : {
        "collider" : { "box" : { "size" : [ 2, 2, 2 ] } }
    }
    "MSFT_RigidBodies" : {
        "physicsMaterial" : 1
    }
}

I'm using the existence of a "physicsMaterial" on a node which has a
collider to indicate that "this is a node which should go into a RB
simulation" - it _feels_ like it would be nice to inherit those
materials, rather than having to specify them on every leaf shape,
however, that might incorrectly cause a "non-RB-simulated" node to be
selected to participate in the simulation, so am inclined to explicitly
say that materials are _not_ inherited.

Alternative might be to add a "default" material to a rigidBody node,
but this adds an additional edge case when importing, and would much
prefer to keep those to a minimum.

REBASE