From discussion; refactoring convex hulls & morph targets
Remove convex hull options from base KHR_collision_shapes; for many
cases, convex hulls are not necessary and having to support convex hull
generation in a viewer may be a significant implementation burden.

Refactor morph targets to support the common use-case of describing a
mesh collider which intends to use the same morph target weights as a
node's instance weights. Add an explicit recommendation for how an
extension can reference a collision shape.